requirements.txt
Flask==3.0.3
Flask-Session==0.5.0
python-dotenv==1.0.1
mysql-connector-python==9.0.0
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5